BookingBot/requirements.txt
aiogram==3.1.1
SQLAlchemy==1.4.46
asyncpg==0.27.0
python-dotenv==1.0.0
aiohttp==3.8.6